master - refactor -
added redux to the application and also the Rredux chrome extension. Modified the index.js file at src to reflect the requirements for the redux debugger to show. extended the js in app.js to have a submitFood function that is tied to the button that passes day, meal, and recipe to the state manager